Changes file structure for backend
